apply format and long in for loops

+++ Boundary/Boundary_test.cpp
@@ -13,7 +13,7 @@ int main() {
     {
         std::cout << "test spherical shell inside" << std::endl;
         std::unique_ptr<Boundary> b = std::make_unique<SphereShell>(center, 2.0, true);
-        for (int i = 0; i < 1000; i++) {
+        for (long i = 0; i < 1000; i++) {
             Evec3 query = (Evec3::Random() * 5) + Emap3(center);
             Evec3 projection = Evec3::Zero();
             Evec3 normI = Evec3::Zero();
@@ -31,7 +31,7 @@ int main() {
     {
         std::cout << "test spherical shell outside" << std::endl;
         std::unique_ptr<Boundary> b = std::make_unique<SphereShell>(center, 2.0, false);
-        for (int i = 0; i < 1000; i++) {
+        for (long i = 0; i < 1000; i++) {
             Evec3 query = (Evec3::Random() * 5) + Emap3(center);
             Evec3 projection = Evec3::Zero();
             Evec3 normI = Evec3::Zero();
@@ -49,7 +49,7 @@ int main() {
     {
         std::cout << "test wall" << std::endl;
         std::unique_ptr<Boundary> b = std::make_unique<Wall>(center, axis);
-        for (int i = 0; i < 1000; i++) {
+        for (long i = 0; i < 1000; i++) {
             Evec3 query = (Evec3::Random() * 5) + Emap3(center);
             Evec3 projection = Evec3::Zero();
             Evec3 normI = Evec3::Zero();
@@ -68,7 +68,7 @@ int main() {
     {
         std::cout << "test tube inside" << std::endl;
         std::unique_ptr<Boundary> b = std::make_unique<Tube>(center, axis, 2.0, true);
-        for (int i = 0; i < 1000; i++) {
+        for (long i = 0; i < 1000; i++) {
             Evec3 query = (Evec3::Random() * 5) + Emap3(center);
             Evec3 projection = Evec3::Zero();
             Evec3 normI = Evec3::Zero();
@@ -86,7 +86,7 @@ int main() {
     {
         std::cout << "test tube outside" << std::endl;
         std::unique_ptr<Boundary> b = std::make_unique<Tube>(center, axis, 2.0, false);
-        for (int i = 0; i < 1000; i++) {
+        for (long i = 0; i < 1000; i++) {
             Evec3 query = (Evec3::Random() * 5) + Emap3(center);
             Evec3 projection = Evec3::Zero();
             Evec3 normI = Evec3::Zero();
