+++ Deprecated/Collision/CPSolver.cpp
@@ -19,9 +19,9 @@ void CPSolver::clipZero(Teuchos::RCP<TV> &vecRcp) const {
     // std::cout << strides[0] << " " << strides[1] << std::endl;
     // std::cout << std::is_same<decltype(x_2d)::array_layout, Kokkos::LayoutLeft>::value << std::endl;
 
-    for (int c = 0; c < x_2d.extent(1); c++) {
+    for (long c = 0; c < x_2d.extent(1); c++) {
 #pragma omp parallel for
-        for (int i = 0; i < x_2d.extent(0); i++) {
+        for (long i = 0; i < x_2d.extent(0); i++) {
             const double temp = x_2d(i, c);
             x_2d(i, c) = temp < 0 ? 0 : temp;
         }
@@ -46,9 +46,9 @@ void CPSolver::maxXY(const Teuchos::RCP<const TV> &vecXRcp, const Teuchos::RCP<c
     auto y_2d = vecYRcp->getLocalView<Kokkos::HostSpace>();
     auto z_2d = vecZRcp->getLocalView<Kokkos::HostSpace>();
     vecZRcp->modify<Kokkos::HostSpace>();
-    for (int c = 0; c < x_2d.extent(1); c++) {
+    for (long c = 0; c < x_2d.extent(1); c++) {
 #pragma omp parallel for
-        for (int i = 0; i < x_2d.extent(0); i++) {
+        for (long i = 0; i < x_2d.extent(0); i++) {
             z_2d(i, c) = std::max(x_2d(i, c), y_2d(i, c));
         }
     }
@@ -72,9 +72,9 @@ void CPSolver::minXY(const Teuchos::RCP<const TV> &vecXRcp, const Teuchos::RCP<c
     auto y_2d = vecYRcp->getLocalView<Kokkos::HostSpace>();
     auto z_2d = vecZRcp->getLocalView<Kokkos::HostSpace>();
     vecZRcp->modify<Kokkos::HostSpace>();
-    for (int c = 0; c < x_2d.extent(1); c++) {
+    for (long c = 0; c < x_2d.extent(1); c++) {
 #pragma omp parallel for
-        for (int i = 0; i < x_2d.extent(0); i++) {
+        for (long i = 0; i < x_2d.extent(0); i++) {
             z_2d(i, c) = std::min(x_2d(i, c), y_2d(i, c));
         }
     }
@@ -99,9 +99,9 @@ double CPSolver::checkResiduePhi(const Teuchos::RCP<const TV> &vecXRcp, const Te
     auto b_2d = vecbRcp->getLocalView<Kokkos::HostSpace>();
     auto z_2d = vecTempRcp->getLocalView<Kokkos::HostSpace>();
     vecTempRcp->modify<Kokkos::HostSpace>();
-    for (int c = 0; c < x_2d.extent(1); c++) {
+    for (long c = 0; c < x_2d.extent(1); c++) {
 #pragma omp parallel for
-        for (int i = 0; i < x_2d.extent(0); i++) {
+        for (long i = 0; i < x_2d.extent(0); i++) {
             z_2d(i, c) = std::min(x_2d(i, c), y_2d(i, c) + b_2d(i, c));
         }
     }
@@ -153,9 +153,9 @@ void CPSolver::hMinMap(const Teuchos::RCP<const TV> &xRcp, const Teuchos::RCP<co
     hRcp->modify<Kokkos::HostSpace>();
     maskRcp->modify<Kokkos::HostSpace>();
 
-    for (int c = 0; c < xView.extent(1); c++) {
+    for (long c = 0; c < xView.extent(1); c++) {
 #pragma omp parallel for
-        for (int i = 0; i < xView.extent(0); i++) {
+        for (long i = 0; i < xView.extent(0); i++) {
             if (xView(i, c) < yView(i, c)) {
                 hView(i, c) = xView(i, c);
                 maskView(i, c) = 1;
@@ -210,8 +210,8 @@ CPSolver::CPSolver(int localSize, double diagonal) {
 
     // a random matrix
     Teuchos::SerialDenseMatrix<int, double> BLocal(localSize, localSize, true); // zeroOut
-    for (int i = 0; i < localSize; i++) {
-        for (int j = 0; j < localSize; j++) {
+    for (long i = 0; i < localSize; i++) {
+        for (long j = 0; j < localSize; j++) {
             BLocal(i, j) = dis(gen);
         }
     }
@@ -219,7 +219,7 @@ CPSolver::CPSolver(int localSize, double diagonal) {
     Teuchos::SerialDenseMatrix<int, double> ALocal(localSize, localSize, true);
     Teuchos::SerialDenseMatrix<int, double> tempLocal(localSize, localSize, true);
     Teuchos::SerialDenseMatrix<int, double> DLocal(localSize, localSize, true);
-    for (int i = 0; i < localSize; i++) {
+    for (long i = 0; i < localSize; i++) {
         DLocal(i, i) = fabs(dis(gen)) + 2;
     }
 
@@ -227,7 +227,7 @@ CPSolver::CPSolver(int localSize, double diagonal) {
     tempLocal.multiply(Teuchos::NO_TRANS, Teuchos::NO_TRANS, 1.0, DLocal, BLocal, 0.0); // temp = DB
     ALocal.multiply(Teuchos::TRANS, Teuchos::NO_TRANS, 1.0, BLocal, tempLocal, 0.0);    // A = B^T DB
 
-    for (int i = 0; i < localSize; i++) {
+    for (long i = 0; i < localSize; i++) {
         ALocal(i, i) += diagonal;
     }
 
@@ -237,9 +237,9 @@ CPSolver::CPSolver(int localSize, double diagonal) {
     double droptol = 1e-7;
     Kokkos::View<size_t *> rowCount("rowCount", localSize);
     Kokkos::View<size_t *> rowPointers("rowPointers", localSize + 1);
-    for (int i = 0; i < localSize; i++) {
+    for (long i = 0; i < localSize; i++) {
         rowCount[i] = 0;
-        for (int j = 0; j < localSize; j++) {
+        for (long j = 0; j < localSize; j++) {
             if (fabs(ALocal(i, j)) > droptol) {
                 rowCount[i]++;
             }
@@ -247,14 +247,14 @@ CPSolver::CPSolver(int localSize, double diagonal) {
     }
 
     rowPointers[0] = 0;
-    for (int i = 1; i < localSize + 1; i++) {
+    for (long i = 1; i < localSize + 1; i++) {
         rowPointers[i] = rowPointers[i - 1] + rowCount[i - 1];
     }
     Kokkos::View<int *> columnIndices("columnIndices", rowPointers[localSize]);
     Kokkos::View<double *> values("values", rowPointers[localSize]);
     int p = 0;
-    for (int i = 0; i < localSize; i++) {
-        for (int j = 0; j < localSize; j++) {
+    for (long i = 0; i < localSize; i++) {
+        for (long j = 0; j < localSize; j++) {
             if (fabs(ALocal(i, j)) > droptol) {
                 columnIndices[p] = j;
                 values[p] = ALocal(i, j);
@@ -267,7 +267,7 @@ CPSolver::CPSolver(int localSize, double diagonal) {
     const int colIndexCount = rowPointers[localSize];
     std::vector<int> colMapIndex(colIndexCount);
 #pragma omp parallel for
-    for (int i = 0; i < colIndexCount; i++) {
+    for (long i = 0; i < colIndexCount; i++) {
         colMapIndex[i] = columnIndices[i] + myRank * localSize;
     }
 
@@ -582,8 +582,8 @@ class mmNewtonOperator : public TOP {
         auto yView = Y.getLocalView<Kokkos::HostSpace>();
         auto maskView = maskRcp->getLocalView<Kokkos::HostSpace>();
         Y.modify<Kokkos::HostSpace>();
-        for (int c = 0; c < xView.extent(1); c++) {
-            for (int i = 0; i < xView.extent(0); i++) {
+        for (long c = 0; c < xView.extent(1); c++) {
+            for (long i = 0; i < xView.extent(0); i++) {
                 if (maskView(i, c) > 0.5) {
                     yView(i, c) = xView(i, c);
                 }
@@ -839,8 +839,8 @@ int CPSolver::test_LCP(double tol, int maxIte, int solverChoice) {
     int yerrorN = 0;
     double yerrormax = 0.0;
     assert(xView.extent(0) == mapRcp->getNodeNumElements());
-    for (int c = 0; c < xView.extent(1); c++) {
-        for (int i = 0; i < xView.extent(0); i++) {
+    for (long c = 0; c < xView.extent(1); c++) {
+        for (long i = 0; i < xView.extent(0); i++) {
             if (xView(i, c) < 0) {
                 xerrorN++;
                 xerrormax = std::max(xerrormax, static_cast<double>(-xView(i, c)));
